requirements.txt
# Database requirements
sqlalchemy==2.0.23
alembic==1.13.0

# Database drivers
psycopg2-binary==2.9.9  # PostgreSQL
# OR use psycopg2 for production (requires compilation)

# Scraping dependencies
recipe-scrapers==14.55.0
requests==2.31.0
urllib3==2.1.0

# CLI
click==8.1.7

# Configuration and validation
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.2.0

# API Framework
fastapi==0.109.0
uvicorn[standard]==0.26.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib==1.7.4
bcrypt==3.2.2

# Development and testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
responses==0.24.1
freezegun==1.4.0
black==23.11.0
mypy==1.7.1
sqlalchemy-stubs==0.4
